+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Codex CLI: Use `RUST_LOG=warning` (rather than `debug`) to reduce amount of output produced by `exec()`.
+- Handle multiple `run()` calls to the same agent (resume session and send only new messages in prompt).
+
 ## 0.2.26 (15 November 2025)
 
 - Copy agent binary executables to /var/tmp/ (more liberal default permissions)